# anything/open-api/generator_sample/out/src/openapi_server/models/address.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401


class Address(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    Address - a model defined in OpenAPI

        street: The street of this Address [Optional].
        city: The city of this Address [Optional].
        state: The state of this Address [Optional].
        zip: The zip of this Address [Optional].
    """

    street: Optional[str] = Field(alias="street", default=None)
    city: Optional[str] = Field(alias="city", default=None)
    state: Optional[str] = Field(alias="state", default=None)
    zip: Optional[str] = Field(alias="zip", default=None)

Address.update_forward_refs()